KXml looks good?
git-svn-id: https://repo.kowalczyk.info/svn/ars_framework/trunk/ars_framework@299 10a9aba9-86da-0310-ac04-a2df2cc00fd9

+++ Src/KXmlParser/KXmlParser.cpp
@@ -200,7 +200,7 @@ void KXmlParser::push(int c)
 {
     isWhitespace_ &= c <= ' ';
 
-    if (txtPos_ == txtBuf_.length()) 
+    if (txtPos_ == (int)txtBuf_.length()) 
         txtBuf_.resize((txtBuf_.size()*4)/3 + 4);
     txtBuf_[txtPos_++] = (char) c;
 }
